requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
drf-yasg==1.21.3
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
pyparsing==3.0.9
pytz==2022.1
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
soupsieve==2.3.2.post1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
